pyproject.toml
[tool.poetry]
name = "genomeplotter"
version = "0.1.0"
description = "A small toolset to generate visualization of the human genome."
authors = ["Daniel Suveges <daniel.suveges@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.15"
pandas = "^1.5.2"
cairosvg = "^2.5.2"
requests = "^2.28.1"
pyspark = "^3.3.1"
pybedtools = "^0.9.0"
hydra-core = "^1.3.1"
types-python-dateutil = "^2.8.19.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
black = "^22.12.0"
mypy = "^0.991"
flake8 = "^6.0.0"
pep8-naming = "^0.13.3"
interrogate = "^1.5.0"
isort = "^5.11.4"
darglint = "^1.8.1"
flake8-docstrings = "^1.6.0"
flake8-annotations = "^2.9.1"
flake8-quotes = "^3.3.2"
flake8-simplify = "^0.19.3"
pytest = "^7.2.0"
flake8-tidy-imports = "^4.8.0"
flake8-type-checking = "^2.3.0"
flake8-typing-imports = "^1.14.0"
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
